Popover: update position even when own props stay the same
Previously, the Popover didn't update its position when its children changed. This would lead to a mis-positioned Popover when the height of its children was changed.

This change makes sure that the Popover recalculates its position even when its props are the same. This makes sure that it is correctly positioned when only its children change. This change uses `this.isUpdatingPosition` to avoid infinite updating loops.