WIP: Rewind: Replace all uses of the old /activity/site/rewind API
We have been polling two APIs for the same information but the old one
was forcing us to duplicate in slightly-different ways much of the logic
already performed on the server.

This is part of a larger project to refactor the Rewind and Activity Log
UIs such that we can rely on the so-called "state machine API" as our
central source of information for all-things Rewind-related.